test: change test directory structure
